Support per-operation consistency levels
This gives more control, e.g. if you want most reads to be eventually
consistent but some to be highly consistent.

There is a new Option field: Consistency.  Two new QueryExecutor
functions take an Options argument and pass that through to gocql
query: QueryWithOptions() and ExecuteWithOptions().  If not set
specified, the default consistency level (currently "One") will be used.