support maps that are not map[string]interface{}
The reflect code in gocassa does not treat maps in any special way when
converting structs to maps for internal use. This means that if your
struct has a map[string]string, it will remain a map[string]string.
However when using MapSetFields the requested changes are in a
map[string]interface{} so we need to ensure that when we copy from
MapSetFields map[string]interface{} to the record value that we don't
assume the map in the record is also a map[string]interface{}, but could
be, for example, a map[string]string.