make sure that mock returns an error for an empty primary key
Currently, we only check whether or not the key is present, we also need
to check if the value is an empty string. As far as I understand, it is
safe to write zero values for other types like int, time etc. in the
primary key, and writes only fail for empty strings.

This is a hack, and I couldn't figure out any other way of solving this.

I really wanted to fix this because real C* behavior would be to return
an error here, while unit tests with mocked keyspaces would happily
pass.