allow any component of a composite primary key to be empty in mocks
The validation for partition key that I previously added did not allow
any part of a composite primary key to be empty, whereas a real C*
cluster allows you to write empty keys without throwing any validation
error.

This behaviour is inconsistent from the validation error thrown when
there's an empty partition key in maptable and multimaptable.

I have an example in this document-
https://paper.dropbox.com/doc/C-partition-key-validation--AZ_wjbcLH2Svw2jYWnjxxfrAAg-rfobnXlarpbJkZIcm58rs

This patch makes sure that the "Missing mandatory PRIMARY KEY part xx"
is only thrown if there's a single partition key of type string which
is empty.