Support custom unmarshaling to decode custom types
User-defined types can implement `gocql.Unmarshaler` interface to supply
their own unmarshal method from CQL fields. Unfortunately `gocql` backend
retrieves fields as `map[string]interface{}`, which ignore any type info
in the retrieved object, so the fields with custom unmarshaler are not
unmarshaled using the custom unmarshal method, and passed to `gocassa` as
an ordinary string. We get decode errors when `gocassa` attempts to decode
those custom `string` fields into the concrete struct object the user provides.

To unmarshal the fields correctly, we need to use the custom unmarshal methods
when reading an object; instead of relying on `gocql` to recognise the string
is a custom field, we use the concrete struct object provided by the user to
`gocassa` as a hint to use the custom marshal methods.

This commit adds a special hook function to decode custom types that implement
`gocql.Unmarshaler`. When decoding a `string` field that implements it, create
a new instance of the field type and decode the data into that instance.