💎 Refactor gocassa code to return a Statement interface
This replaces quite a bit of the gocassa internals to return an object
for the query and values rather than a string and a variadic interface.

This refactoring is for down the line so we can return also the Column
Names in the Statement and potentially more data in the future if
needed. To achieve this, we've had to tweak the public interface too
significantly hence the bump to 2.0.