🏂 Add field types to the Statement interface
This PR adds type information to the fields we are selecting if making
a SELECT query. This information can be used by consumers to do
reflection and allocate the relevant pointers for scanning using the
gocql iter.Scan method.

To be frankly honest, it's quite a leaky abstraction however it is super
optimal because you don't end up doing multiple reflection steps and can
support things like custom CQL Unmarshal methods without a lot of work.