Merge existing options with mocking WithOptions (#48)
The behaviour with Op's in normal usage is to merge any Options specified
in WithOptions with the Op's existing Options. This was not being
followed in mockOp. This commit fixes this by making sure we always merge
with any existing Options, thus preventing the mocked WithOptions
unsetting any Options already set.

This commit also adds a test case to cover this.