🖨️ Add a Scanner/Scannable interface to be exposed
Initially I wanted to seperate the concerns of iteration to the backend
so it had more control. However, this would mean each backend would need
to do reflection business which is hard to test for and hard to assert
for correctness. It also is a very leaky abstraction because gocassa
would need to expose the field types externally. This is messy and prone
to error.

Instead, this change passes in a scanner which takes an iterator. We can
hook in a mock iterator (as long as it passes the same interface) or
a gocql iterator. That way, gocassa is in control of what happens with
the iterator for scanning rows.

In order for the Scanner to work, we've added a bit more reflection code
which will use the cache to ensure that for a given struct type, we only
need to do the hardcore work of figuring out fields once.