🐔 Statement shouldn't have a list of FieldTypes exposed
Similar to the previous commit, we don't want a leaky abstraction around
field types when parsing a statement. Also it turns out a lot of
programs pass in a completely different struct types during unmarshaling
(sometimes with fields being removed or with different types) than the
ones passed in during table initialisation. We could've forbidden this
using a type check as an optimisation (and that option is still there)
but for now it seems quite heavy handed.

Because of this, we can get rid of some cruft we added around logging
the struct on table creation and generating field types.