🏥 Remove the mapstructure library usage as part of the op
This removes the use of the mapstructure library. Now the decode step is
embedded as part of the scanner which handles all sorts of pointer magic
and putting the pointers into the right places in the struct.

Also cleans up a stray use of generateFieldTypes which we got rid of in
a prior commit.