🏍️ Add some tests for our mockIterator Scanner
In an ideal world, we would rip out all of this code and replace it with
a call to UnmarshalCQL with some crafted params. This will give us the
benefit of all the checks that gocql has built in.

Either way, this PR locks down the current behaviour so when we do the
migration in the future, we can assert for major changes.