💂 Clean up semantics around reflect field types and add tests
Part of this work involved some new functions in the reflect side of
things to take a struct and turn it into a map. We've also changed the
semantics of Field (especially now that it's public) to not
automatically follow pointers. If a struct type is *int, that shouldn't
be represented as int as the type.