💄 Clean up scanner and add tests
This commit is the cleanup commit. We've all been there, you've been
prototyping and in the middle of getting things done, you write a bunch
of hacky code all over the place. Well for this refactoring, this was
indeed the scanner. There was a bunch of duplication for reading into
a slice of structs or reading into a single struct. All of this has been
cleaned up and refactored into nice logical functions so you can see
what's happening.

As a result, we've been able to add a bunch of tests for all the edge
cases encountered so far. This boosts the test coverage of the scanner
which is great because it's one of the more fragile components with all
the reflection it needs to do.

I've also taken this opportunity to rename the ScanAll to ScanIter. My
rationale for this is because ScanAll feels like it's going to scan
everything in the iterator. This might not be the case all the time
especially if you pass in a single struct where multiple results have
been returned.