🕸️ Support map types in mock
Map types are a bit unique in the mock because in order to dynamically
do diff based upserts, the map is treated like an interface map value
type ( map[<KeyType>]interface{}) in certain cases. In order to support
this, we do some type assertions in the map case.

Also fixes a test in mock_test.go which we changed previously to
accomodate the refactor. This means we haven't had to change any of the
mock behaviour with this refactoring and all existing use cases work!