🐞 Skip over zero values (values not set) in our mocks
If a value hasn't been set previously and we try and select it, it might
be the zero value and that'll cause a panic if we call Type() or similar
on reflect.Zero. This commit adds a valid check to ensure we don't have
the zero type (which is something we can just continue over).