👔 Take in a gocql Scannable to run our scanning
We need to take in a gocql like Scannable to really get at the errors
within gocassa. This is so we can properly make the error handling
between the iterator and the gocassa scanner handle RowNotFound
properly.

The problem we have is calling iter.Scan can just return a boolean and
gocassa used to assume that this would mean there was a RowNotFound.
However, this is not always the case because the iterator could just
return false if it has a general error.