Rename RunAtomically to RunLoggedBatch
The name RunAtomically is misleading - it doesn't give atomicity in the
"traditional" sense. It doesn't provide isolation (so other readers may
observe inconsistent state), and it doesn't provide Read-Your-Writes
consistency. That is to say, assuming the following three operations are
the _only_ operations issued on the database:

 * A write with RunAtomically (which returns an error)
 * A read of the written row
 * A second read of the written row

In some failure cases, both reads may return different values (in
particular, the first read can observe the old value and the second read
the new value), even if all operations were to be performed at QUORUM
consistency

The name RunAtomically is therefore an attractive nuisance. The name
RunLoggedBatch - which takes its name from the underlying Cassandra
operation - means someone is much more likely to peruse the
documentation before attempting to use this Cassandra feature